added exception handling

+++ src/main/java/com/lucasnunes/webservices/services/UserService.java
@@ -2,7 +2,11 @@ package com.lucasnunes.webservices.services;
 
 import com.lucasnunes.webservices.entities.User;
 import com.lucasnunes.webservices.repositories.UserRepository;
+import com.lucasnunes.webservices.services.exceptions.DatabaseException;
+import com.lucasnunes.webservices.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -18,23 +22,34 @@ public class UserService {
         return repository.findAll();
     }
 
-    public User findById(Long id){
+    public User findById(Long id) {
         Optional<User> obj = repository.findById(id);
-        return obj.get();
+        return obj.orElseThrow(() -> new ResourceNotFoundException(id));
     }
 
-    public User insert (User obj) {
+    public User insert(User obj) {
         return repository.save(obj);
     }
 
-    public void delete (Long id){
-        repository.deleteById(id);
+    public void delete(Long id) {
+        try {
+            if (!repository.existsById(id)) {
+                throw new ResourceNotFoundException(id);
+            }
+            repository.deleteById(id);
+        } catch (DataIntegrityViolationException e) {
+            throw new DatabaseException(e.getMessage());
+        }
     }
 
-    public User update(Long id, User obj){
-        User entity = repository.getReferenceById(id);
-        updateData(entity, obj);
-        return repository.save(entity);
+    public User update(Long id, User obj) {
+        try {
+            User entity = repository.getReferenceById(id);
+            updateData(entity, obj);
+            return repository.save(entity);
+        } catch (EntityNotFoundException e) {
+            throw new ResourceNotFoundException(id);
+        }
     }
 
     private void updateData(User entity, User obj) {
